Raise MSRV to 1.89; take ignore, fastcdc 5, drop fs4; bump actions
- rust-version 1.85 -> 1.89. ignore 0.4.30+ needs 1.88 (let chains) and
  fastcdc 5 needs 1.87; 1.89 also brings std File::lock/unlock, so fs4 is
  dropped in favour of the standard library.
- fastcdc 3.2 -> 5.0: chunk sizes are usize now. Cut points are unchanged
  for our power-of-two sizes (verified: identical chunk maps from both
  versions on a 50 MB file and an odd-length file), so existing stores and
  remotes keep deduplicating.
- Apply clippy fixes that the higher MSRV unlocks (let chains,
  is_multiple_of).
- CI: add an msrv job that builds on the rust-version from Cargo.toml.
- Actions: setup-uv v7, upload-artifact v7, download-artifact v8,
  action-gh-release v3 (Node 24 runtimes). Applied directly because the gh
  token lacks the workflow scope needed to merge Dependabot's PRs.

Supersedes #2, #3, #4, #5, #6, #8, #9.

+++ crates/pollard-objects/src/store.rs
@@ -12,9 +12,9 @@ use crate::{Entry, Error, IoCtx, Manifest, Result, Snapshot, WalkOptions, hash_b
 
 /// Blobs under this size are stored whole in `objects/`; larger ones are CDC-chunked.
 pub const SMALL_LIMIT: u64 = 1 << 20;
-const CHUNK_MIN: u32 = 8 * 1024;
-const CHUNK_AVG: u32 = 64 * 1024;
-const CHUNK_MAX: u32 = 128 * 1024;
+const CHUNK_MIN: usize = 8 * 1024;
+const CHUNK_AVG: usize = 64 * 1024;
+const CHUNK_MAX: usize = 128 * 1024;
 const ZSTD_LEVEL: i32 = 3;
 
 /// On-disk store rooted at `.pollard/`:
